Rebase cleanup
